# Conflicts:
#	src/main/resources/META-INF/sql/schema.sql

+++ src/main/java/com/manager/pulsar/entity/ClustersEntity.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 public class ClustersEntity {
-    private int clusterId;
+    private long clusterId;
     private String cluster;
 
     private String serviceUrl;
